Move content to wikipedia/utils.py
