--- concerned-coyotes/earlyinternet/utils/helpers.py
@@ -1,9 +1,0 @@
-def create_url_with_query_parameters(base_url: str, params: dict) -> str:
-    """Creates a url for the given base address with given parameters
-    as a query string."""
-
-    parameter_string = "&".join(f"{key}={value}"
-                                for key, value in params.items())
-    url = f"{base_url}?{parameter_string}"
-
-    return url
