+++ concerned-coyotes/earlyinternet/wikipedia/utils.py
@@ -6,8 +6,6 @@ from time import mktime
 
 import feedparser
 
-from .helpers import create_url_with_query_parameters
-
 
 class WikipediaFeaturedArticleParser(HTMLParser):
     def __init__(self) -> "WikipediaFeaturedArticleParser":
@@ -103,3 +101,14 @@ def get_wikipedia_featured_articles() -> typing.List[dict]:
         })
 
     return articles
+
+
+def create_url_with_query_parameters(base_url: str, params: dict) -> str:
+    """Creates a url for the given base address with given parameters
+    as a query string."""
+
+    parameter_string = "&".join(f"{key}={value}"
+                                for key, value in params.items())
+    url = f"{base_url}?{parameter_string}"
+
+    return url
